kpb: fix logging cast issue
We cannot assume the build will always be 32bits, therefore when casting
an int to a ptr use uintptr_t

+++ src/audio/kpb.c
@@ -522,7 +522,7 @@ static int kpb_prepare(struct comp_dev *dev)
 
 	if (!kpb->sel_sink || !kpb->host_sink) {
 		comp_info(dev, "kpb_prepare(): could not find sinks: sel_sink %d host_sink %d",
-			  (uint32_t)kpb->sel_sink, (uint32_t)kpb->host_sink);
+			  (uintptr_t)kpb->sel_sink, (uintptr_t)kpb->host_sink);
 		ret = -EIO;
 	}
